enhancement: add caller context
